feat(toc): ✨ Add TOC options and scoping modes
* Introduced `TocOptions` class to control TOC generation settings.
* Added `TocScope` enum for defining scoping modes for TOC.
* Enhances flexibility in generating tables of contents in Markdown documents.
